+++ react/AddressRules.test.js
@@ -3,6 +3,7 @@ import AddressRules from './AddressRules'
 import { shallow, render, waitForElement } from 'test-utils'
 import defaultRules from './country/default'
 import braRules from './country/BRA'
+import { getField } from './selectors/fields'
 
 describe('AddressRules', () => {
   it('should load the defined rules', async () => {
@@ -31,16 +32,16 @@ describe('AddressRules', () => {
       </AddressRules>,
     )
 
-    const result = await waitForElement(
-      () => getByTestId(testId),
-      { container }
-    )
+    const result = await waitForElement(() => getByTestId(testId), {
+      container,
+    })
 
     expect(result).toBeDefined()
   })
 
   it('should provide default rules when country is unrecognized', async () => {
-    global.console = { warn: jest.fn() }
+    const prevWarn = global.console.warn
+    global.console.warn = jest.fn()
 
     const instance = shallow(
       <AddressRules
@@ -52,6 +53,37 @@ describe('AddressRules', () => {
     ).instance()
 
     const rules = await instance.componentDidMount()
+
     expect(rules).toEqual(defaultRules)
+
+    global.console.warn = prevWarn
+  })
+
+  it('should merge geolocation field rules with default field rules', async () => {
+    const instance = shallow(
+      <AddressRules
+        country={'BRA'}
+        fetch={country => import('./country/' + country)}
+        useGeolocation
+      >
+        <h1>It works!</h1>
+      </AddressRules>,
+    ).instance()
+
+    const { default: initialRules } = await import('./country/BRA')
+    const rules = await instance.componentDidMount()
+
+    expect(getField('postalCode', initialRules)).toMatchObject({
+      required: true,
+    })
+    expect(getField('postalCode', rules)).toMatchObject({
+      required: false,
+    })
+
+    expect(getField('number', initialRules).notApplicable).toBeUndefined()
+    expect(getField('number', rules)).toMatchObject({
+      required: true,
+      notApplicable: true,
+    })
   })
 })