+++ react/country/CAN.js
@@ -104,8 +104,13 @@ export default {
       types: ['postal_code'],
       required: false,
     },
-    number: { valueIn: 'long_name', types: ['street_number'], required: false, notApplicable: true },
-    street: { valueIn: 'long_name', types: ['route'], required: false },
+    number: {
+      valueIn: 'long_name',
+      types: ['street_number'],
+      required: true,
+      notApplicable: true,
+    },
+    street: { valueIn: 'long_name', types: ['route'] },
     neighborhood: {
       valueIn: 'long_name',
       types: [
@@ -116,45 +121,26 @@ export default {
         'sublocality_level_4',
         'sublocality_level_5',
       ],
-      required: false,
     },
     state: {
       valueIn: 'short_name',
       types: ['administrative_area_level_1'],
-      required: false,
     },
     city: {
       valueIn: 'long_name',
       types: ['administrative_area_level_2', 'locality'],
-      required: false,
     },
   },
   summary: [
     [
-      {
-        name: 'complement',
-        delimiterAfter: '-',
-      },
-      {
-        name: 'number',
-      },
-      {
-        delimiter: ' ',
-        name: 'street',
-      },
+      { name: 'complement', delimiterAfter: '-' },
+      { name: 'number' },
+      { delimiter: ' ', name: 'street' },
     ],
     [
-      {
-        name: 'city',
-      },
-      {
-        delimiter: ' ',
-        name: 'state',
-      },
-      {
-        delimiter: '  ',
-        name: 'postalCode',
-      },
+      { name: 'city' },
+      { delimiter: ' ', name: 'state' },
+      { delimiter: '  ', name: 'postalCode' },
     ],
   ],
 }