+++ react/country/GTM.js
@@ -673,57 +673,34 @@ export default {
     number: {
       valueIn: 'long_name',
       types: ['street_number'],
-      required: false,
+      required: true,
       notApplicable: true,
     },
-    street: { valueIn: 'long_name', types: ['route'], required: false },
+    street: { valueIn: 'long_name', types: ['route'] },
     neighborhood: {
       valueIn: 'long_name',
       types: ['neighborhood'],
-      required: false,
     },
     state: {
       valueIn: 'long_name',
       types: ['administrative_area_level_1'],
-      required: false,
     },
     city: {
       valueIn: 'long_name',
       types: ['administrative_area_level_2', 'locality'],
-      required: false,
     },
   },
   summary: [
     [
-      {
-        name: 'street',
-      },
-      {
-        delimiter: ' ',
-        name: 'number',
-      },
-      {
-        delimiter: ', ',
-        name: 'complement',
-      },
+      { name: 'street' },
+      { delimiter: ' ', name: 'number' },
+      { delimiter: ', ', name: 'complement' },
     ],
     [
-      {
-        name: 'neighborhood',
-        delimiterAfter: ' - ',
-      },
-      {
-        name: 'city',
-      },
-      {
-        delimiter: ' - ',
-        name: 'state',
-      },
+      { name: 'neighborhood', delimiterAfter: ' - ' },
+      { name: 'city' },
+      { delimiter: ' - ', name: 'state' },
     ],
-    [
-      {
-        name: 'postalCode',
-      },
-    ],
+    [{ name: 'postalCode' }],
   ],
 }