+++ react/country/ROU.js
@@ -13991,19 +13991,17 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
     },
     number: {
       valueIn: 'long_name',
       types: ['street_number'],
-      required: false,
+      required: true,
       notApplicable: true,
     },
-    street: { valueIn: 'long_name', types: ['route'], required: true },
+    street: { valueIn: 'long_name', types: ['route'] },
     state: {
       valueIn: 'long_name',
       types: ['administrative_area_level_1'],
-      required: true,
       handler: address => {
         if (!address.city || !address.state) {
           return address
@@ -14034,7 +14032,6 @@ export default {
     city: {
       valueIn: 'long_name',
       types: ['locality', 'administrative_area_level_2'],
-      required: true,
       handler: (address, googleAddress, pass) => {
         if (address.state && address.state.value === 'Bucuresti') {
           address.city.value = 'Bucuresti'
@@ -14061,40 +14058,19 @@ export default {
     neighborhood: {
       valueIn: 'long_name',
       types: ['administrative_area_level_3'],
-      required: false,
     },
   },
   summary: [
     [
-      {
-        name: 'street',
-      },
-      {
-        delimiter: ' ',
-        name: 'number',
-      },
-      {
-        delimiter: ' ',
-        name: 'complement',
-      },
+      { name: 'street' },
+      { delimiter: ' ', name: 'number' },
+      { delimiter: ' ', name: 'complement' },
     ],
     [
-      {
-        name: 'neighborhood',
-        delimiterAfter: ' - ',
-      },
-      {
-        name: 'city',
-      },
-      {
-        delimiter: ' - ',
-        name: 'state',
-      },
+      { name: 'neighborhood', delimiterAfter: ' - ' },
+      { name: 'city' },
+      { delimiter: ' - ', name: 'state' },
     ],
-    [
-      {
-        name: 'postalCode',
-      },
-    ],
+    [{ name: 'postalCode' }],
   ],
 }