+++ react/country/URY.js
@@ -430,9 +430,8 @@ export default {
     complement: {
       valueIn: 'long_name',
       types: ['street_number'],
-      required: false,
     },
-    street: { valueIn: 'long_name', types: ['route'], required: false },
+    street: { valueIn: 'long_name', types: ['route'] },
     neighborhood: {
       valueIn: 'long_name',
       types: [
@@ -443,50 +442,27 @@ export default {
         'sublocality_level_4',
         'sublocality_level_5',
       ],
-      required: false,
     },
     state: {
       valueIn: 'long_name',
       types: ['administrative_area_level_1'],
-      required: false,
     },
     city: {
       valueIn: 'long_name',
       types: ['administrative_area_level_2', 'locality'],
-      required: false,
     },
   },
   summary: [
     [
-      {
-        name: 'street',
-      },
-      {
-        delimiter: ' ',
-        name: 'number',
-      },
-      {
-        delimiter: ' ',
-        name: 'complement',
-      },
+      { name: 'street' },
+      { delimiter: ' ', name: 'number' },
+      { delimiter: ' ', name: 'complement' },
     ],
+    [{ name: 'neighborhood' }],
     [
-      {
-        name: 'neighborhood',
-      },
-    ],
-    [
-      {
-        name: 'postalCode',
-      },
-      {
-        delimiter: ' - ',
-        name: 'city',
-      },
-      {
-        delimiter: ', ',
-        name: 'state',
-      },
+      { name: 'postalCode' },
+      { delimiter: ' - ', name: 'city' },
+      { delimiter: ', ', name: 'state' },
     ],
   ],
 }